requirements.txt
streamlit==0.80.0
pandas==1.1.3
numpy==1.20.3
Pillow==8.2.0
